Fix an error causing 100% CPU usage and no detection of empty lab.
When there was no devices left on premises dnsmasq no longer produces
any rows it causes read timeout to occur on is_data_available().
That should be normal.

However, the timeout timestamp is not updated and the old timestamp
from the last known journald entry is used when determining the next
timeout timestamp (last journald entry date + $dhch_lease_secs).
If there was a timeout, we should never it but take the current
timestamp from the system clock.

That allows the leave detection to work because $current_time is used
to get current visitors from the database. After that it continues to
the upper part of the loop where infinite wait is triggered.

That way it reports the lab is empty and then waits forever for the
next event which happens when someone enters the premises or any
unknown device causes DHCP activity on the journal.

The bug was there for two years but never triggered because we had a
WLAN access point on our old site. It was configured to use DHCP which
caused regular activity on dnsmasq journal. On our new site we have
all our infrastructure devices on management VLAN and they have static
IP addresses which caused the log to be really quiet when there was
nobody in.

Thanks for reading and have a good day. :-)